# food-platform/services/auth/openapi/auth.yaml
openapi: 3.1.0
info:
  title: Auth Service API
  version: 1.0.0
  description: |
    Authentication service for the Food Delivery Platform.
    Handles phone+OTP login, JWT issuance, refresh token rotation, and logout.

    ## Authentication Flow
    1. Client sends phone number to `/auth/otp/send`
    2. Server sends OTP via SMS
    3. Client sends OTP code to `/auth/otp/verify`
    4. Server returns access token + refresh token
    5. Client uses access token for API calls
    6. When access token expires, client uses refresh token to get new one

  contact:
    name: Food Platform Team
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.staging.food-platform.com
    description: Staging
  - url: https://api.food-platform.com
    description: Production

paths:
  /health:
    get:
      summary: Health check
      description: Returns the health status of the service
      operationId: health
      tags: [Health]
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ready:
    get:
      summary: Readiness check
      description: Returns the readiness status (checks dependencies)
      operationId: ready
      tags: [Health]
      security: []
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/auth/otp/send:
    post:
      summary: Send OTP
      description: Sends a one-time password to the given phone number via SMS
      operationId: sendOtp
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendOtpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/auth/otp/verify:
    post:
      summary: Verify OTP
      description: |
        Verifies the OTP code and returns access + refresh tokens.
        Creates a new user if the phone number is new (for customer/driver roles).
      operationId: verifyOtp
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access token.
        The old refresh token is invalidated (rotation).
      operationId: refreshToken
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/logout:
    post:
      summary: Logout
      description: |
        Revokes the user's session(s).
        If X-Session-ID header is provided, only that session is revoked.
        Otherwise, all sessions for the user are revoked.
      operationId: logout
      tags: [Auth]
      security:
        - BearerAuth: []
      parameters:
        - name: X-Session-ID
          in: header
          description: Optional session ID to revoke (if not provided, all sessions are revoked)
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: logged_out
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SendOtpRequest:
      type: object
      required:
        - phone
        - role
      properties:
        phone:
          type: string
          description: Egyptian phone number (11 digits starting with 01, or with +20/20 prefix)
          example: "01012345678"
          pattern: '^(\+20|20)?01[0-2,5][0-9]{8}$'
        role:
          type: string
          enum: [customer, driver, restaurant, support_l1, support_l2, ops_manager, finance, super_admin, field_supervisor, hr, read_only_analyst]
          description: The role of the user requesting OTP
          example: customer

    SendOtpResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
          description: UUID of the OTP request (use this to verify)
          example: "550e8400-e29b-41d4-a716-446655440000"
        expires_in:
          type: integer
          description: OTP expiry time in seconds
          example: 120
        attempts_remaining:
          type: integer
          description: Maximum verification attempts allowed
          example: 3

    VerifyOtpRequest:
      type: object
      required:
        - request_id
        - code
      properties:
        request_id:
          type: string
          format: uuid
          description: The request ID returned from /auth/otp/send
          example: "550e8400-e29b-41d4-a716-446655440000"
        code:
          type: string
          description: 6-digit OTP code
          example: "123456"
          pattern: '^[0-9]{6}$'
        device_fingerprint:
          type: string
          description: Device fingerprint for session binding (optional, can also be sent via X-Device-Fingerprint header)
          example: "hash-of-device-features"

    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: The refresh token received during authentication
          example: "a1b2c3d4e5f6..."

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (15 min TTL)
          example: "eyJhbGciOiJIUzI1NiIs..."
        refresh_token:
          type: string
          description: Refresh token (30 day TTL, one-time use)
          example: "a1b2c3d4e5f6..."
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
        token_type:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        phone:
          type: string
          example: "01012345678"
        email:
          type: string
          format: email
          nullable: true
          example: "ahmed@example.com"
        name:
          type: string
          example: "Ahmed Mohamed"
        role:
          type: string
          example: customer
        trust_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 92

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: auth
        version:
          type: string
          example: 1.0.0

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_INVALID_OTP
            message:
              type: string
              example: Invalid or expired OTP
            details:
              type: object
              additionalProperties: true

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Health
    description: Health check endpoints